Rename db_keybase to legacy.
Remove duplicate tests
